create first servicefor Grade

+++ src/main/java/edu/itstep/academy/repository/GradeRepository.java
@@ -6,6 +6,6 @@ import java.util.List;
 public interface GradeRepository {
     List<Grade> getAll();
     void saveOrUpdate(Grade grade);
-    void deleteById(int id);
-    Grade getById(int id);
+    void deleteById(Long id);
+    Grade getById(Long id);
 }
